Overhaul. Use Requests for HTTP, remove dodgy and broken threads, and request multiple channels at once.
The last update of the original was 2014 - let's update things, shall we?

This fixes a lot of the issues with the old version, queries multiple channels, rather than just one at a time, and updates the channel topic (Which is otherwise unused) with the current status of the stream. Also eats USERSTATE messages.